Refactoring inputs
