+++ src/components/InputBlocks/index.js
@@ -1,4 +1,22 @@
 // UPLOAD INPUT BLOCK
 import UploadInputBlock from './UploadInputBlock';
 
-export { UploadInputBlock };
+// TOGGLE INPUT BLOCK
+import ToggleInputBlock from './ToggleInputBlock';
+
+// NUMBER INPUT BLOCK
+import NumberInputBlock from './NumberInputBlock';
+
+// SELECT INPUT BLOCK
+import SelectInputBlock from './SelectInputBlock';
+
+// TEXT INPUT BLOCK
+import TextInputBlock from './TextInputBlock';
+
+export {
+  SelectInputBlock,
+  UploadInputBlock,
+  ToggleInputBlock,
+  NumberInputBlock,
+  TextInputBlock,
+};
